Many changes, bug fixes in MMD kernels and generated MMD 3d
Fixed bug in vglIpl4to3Channels where the operator set the image channel
to 4, even tho its 3.

Fixed bug in MMD kernels where the value of the convolution window were
stored in a int but the value was float, causing loss of data.

Fixed Benchmark_MMD to save output image for every operand correctly.

Generated vglCl3dFuzzyXXXDilate and erode for every type of fuzzy
morphological operator.

+++ src/CL_MM/vglClFuzzyAlgDilate.cl
@@ -26,7 +26,7 @@ __kernel void vglClFuzzyAlgDilate(__read_only image2d_t img_input,
 		for(int j = -factory; j <= factory; j++)
 		{
 			float4 a = read_imagef(img_input, smp, (int2)(coords.x + i,coords.y + j));
-			int b = convolution_window[conv_controller];
+			float b = convolution_window[conv_controller];
 			float4 S = a*b;
 			pmax = max(pmax,S);
 			conv_controller++;
