+++ src/CL_MM/vglClFuzzyArithDilate.cl
@@ -26,7 +26,7 @@ __kernel void vglClFuzzyArithDilate(__read_only image2d_t img_input,
 		for(int j = -factory; j <= factory; j++)
 		{
 			float4 a = read_imagef(img_input, smp, (int2)(coords.x + i,coords.y + j));
-			int b = convolution_window[conv_controller];
+			float b = convolution_window[conv_controller];
 			float4 S = sqrt(min(a,b)*((a+b)/2));
 			pmax = max(pmax,S);
 			conv_controller++;
