+++ src/CL_MM/vglClFuzzyDrasticErode.cl
@@ -26,7 +26,7 @@ __kernel void vglClFuzzyDrasticErode(__read_only image2d_t img_input,
 		for(int j = -factory; j <= factory; j++)
 		{
 			float4 a = read_imagef(img_input, smp, (int2)(coords.x + i,coords.y + j));
-			int b = 1 - convolution_window[conv_controller]; //complement of mask
+			float b = 1 - convolution_window[conv_controller]; //complement of mask
 			float4 S;
 			if (b == 0)
 				S = a;
