+++ src/CL_MM/vglClFuzzyGeoErode.cl
@@ -26,7 +26,7 @@ __kernel void vglClFuzzyGeoErode(__read_only image2d_t img_input,
 		for(int j = -factory; j <= factory; j++)
 		{
 			float4 a = read_imagef(img_input, smp, (int2)(coords.x + i,coords.y + j));
-			int b = 1 - convolution_window[conv_controller];
+			float b = 1 - convolution_window[conv_controller];
 			float4 S = 1 - sqrt((1-a)*(1-b));
 			pmin = min(pmin,S);
 			conv_controller++;
